requirements.txt
openai-whisper
google-generativeai